changed module names to lowercase
