+++ WordRPG/states/Game.py
@@ -1,7 +1,8 @@
 """ Placeholder state for main 'game' loop """
 from collections import deque
-from ..gui.Screen import const, Screen, setup_terminal
-from .States import State
+
+from ..gui.screen import const, Screen, setup_terminal
+from .states import State
