+++ WordRPG/states/Main_Menu.py
@@ -1,7 +1,7 @@
 """ 'main_menu' state """
 
-from ..gui.Screen import const, Screen, setup_terminal
-from .States import State
+from ..gui.screen import const, Screen, setup_terminal
+from .states import State
