+++ WordRPG/states/New_Game.py
@@ -1,7 +1,7 @@
 """ 'new_game' state.  Includes character creation. """
 
-from ..gui.Screen import const, Screen, setup_terminal
-from .States import State
+from ..gui.screen import const, Screen, setup_terminal
+from .states import State
