+++ WordRPG/states/Quit.py
@@ -1,4 +1,4 @@
-from .States import Confirm
+from .states import Confirm
