+++ WordRPG/states/States.py
@@ -16,7 +16,7 @@ from copy import deepcopy
 # https://pypi.org/project/keyboard/
 import keyboard
 
-from ..gui.Screen import const, Screen, setup_terminal
+from ..gui.screen import const, Screen, setup_terminal
