+++ WordRPG/states/__init__.py
@@ -1,1 +1,1 @@
-from . import States, FSM
+from . import states, machine
