+++ WordRPG/states/load.py
@@ -1,14 +1,14 @@
 """ 'load_game' state """
 
-from ..gui.Screen import const, Screen, setup_terminal
-from .States import State
+from ..gui.screen import const, Screen, setup_terminal
+from .states import State
 
 
 
-class Save_Game(State):
+class Load_Game(State):
     def __init__(self):
         """ Initiailize class and super class """
-        super(Save_Game, self).__init__()
+        super(Load_Game, self).__init__()
 
         self.screen = self._init_screen()
 
@@ -17,7 +17,7 @@ class Save_Game(State):
         screen = Screen()
         screen.load_screen('scroll', offset=('center',1), fgcolor='YELLOW')
 
-        screen.add_string_to_screen('SAVE GAME', offset=('center', 2),
+        screen.add_string_to_screen('LOAD GAME', offset=('center', 2),
                     fgcolor='CYAN')
 
         return screen
@@ -31,5 +31,5 @@ class Save_Game(State):
     def on_event(self, event, prev_state):
         """ Handles events that are delegated to this State. """
         self.update_screen()
-        self.wait_for_keypress()
+        self.wait_for_keypress()     
         return 'game'   #prev_state
